backend/requirements.txt
click==8.1.3
Flask==2.2.5
Flask-Cors==4.0.1
itsdangerous==2.1.2
Jinja2==3.1.4
MarkupSafe==2.1.1
pymongo==4.6.3
six==1.16.0
Werkzeug==3.0.3
pytest~=7.2.0
gunicorn
kubernetes~=26.1.0
python-dotenv~=0.21.0
PyYAML~=6.0
celery==5.5.3
redis~=4.5, !=4.5.5
ruamel.yaml===0.17.32
PyJWT~=2.9
argon2-cffi~=23.1
Flask-Limiter~=3.8